build(app): improve bundle
